+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 -   migrated to `pyproject.toml`
 -   fixed footer being off-center (#24) (@wroyca)
 -   fixed redundant `auto` in trailing return types (#26) (@wroyca)
+-   added config option `sponsor`
+-   added config option `twitter`
 
 ## v0.12.7 - 2023-07-27